+++ source/colorDBL.cpp
@@ -1,13 +1,13 @@
 #include "../include/colorDBL.h"
 
-void ColorDBL::changeColor(const double& newRed, const double& newGreen, const double& newBlue){
+void ColorDBL::changeColor(const float& newRed, const float& newGreen, const float& newBlue){
     r = newRed;
     g = newGreen;
     b = newBlue;
 }
 
-std::vector<double> ColorDBL::getColor(){
-    return  std::vector<double>{r,g,b};
+std::vector<float> ColorDBL::getColor(){
+    return  std::vector<float>{r,g,b};
 }
 
 // += operator for adding colors
